Hide led/bootstrap-related properties from bootstrapped executable.
The led_builder_is_bootstrapped property is a property that is read by
led to modify its behavior. When it's set, instead of updating the
executable led edit-recipe-bundle will set the led_cas_recipe_bundle
property. In addition to updating the property, if the
led_builder_is_bootstrapped is set when using led edit to modify or
remove properties, it will also set the led_edited_properties and/or
led_removed_properties as appropriate. This allows the bootstrapper to
apply the modified values after applying the properties file. If the
recipe bundle is edited, this also results in led_cas_recipe_bundle
showing up in led_edited_properties.

Since these properties are all meant for managing the interactions
between led and a bootstrapper program, there's no need for those
properties to be sent along to the bootstrapped executable. This updates
the bootstrapper so that those values will be removed from the
properties before triggering the bootstrapped executable.

Change-Id: I694b307784621cabba7c8e263fe7c3cd9683119b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5381578
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#63902}