kron: fix staging limit bug
This bug was causing no builds to be ran on staging. Because we were not
exiting the out loop once we hit the limit we were adding empty items to
the request map. This in turn was leading to panics when it tried
searching for an item in the map.

BUG=None
TEST=local golang debugger

Change-Id: I0df9a2867d87c71e8377edf77f1980e30d7a3623
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5383677
Auto-Submit: Julio Hurtado <juahurta@google.com>
Reviewed-by: Azizur Rahman <azrahman@google.com>
Commit-Queue: Azizur Rahman <azrahman@google.com>
Cr-Commit-Position: refs/heads/main@{#63925}