Fix the make target so that the output isn't clobbered.
The install target depends on PGO if that's enabled, so make sure to
use the default target rather than just "python" when we are doing
a parallel make prior to install -- otherwise it will not be a PGO
build and will just be deleted and rebuilt (serially) by "install".
Disable PGO when cross-compiling.

This gets rid of a lot of wasted work done during the build.

Note that both the "clean" and "install" steps seem to have race
conditions that can cause intermittent failures, so run only the build
step with parallel make.

Change-Id: I462b645b3801412c295fe259c5183e3e1c8ad749
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5367059
Commit-Queue: Brian Ryner <bryner@google.com>
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#63936}