Make completely sure libcrypt is not used during the build.
There is no consistent version of libcrypt available across
the fleet, plus the relevant module is deprecated and unused,
so we want to disable this completely.

Currently autoconf is still detecting these functions, we are
just getting lucky that -lcrypt is not ending up in the final link.

Change-Id: If6a7b4dcb1eb3d46df20757afbfc7b61601b5553
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5376956
Commit-Queue: Brian Ryner <bryner@google.com>
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#63937}