[golangbuild] specify .exe for tools explicitly
For unknown reasons, and correlated with an upgrade of LUCI tools to Go
1.22.1 from Go 1.21.8, it appears that invoking those tools does not
respect PATHEXT. This change specifies .exe for tools explicitly, which
appears to fix the build. I used led to try out the fix, and it seems to
work.

Change-Id: Ic6da5d954738fb7f5736be27a92f7e33311b3180
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5404271
Auto-Submit: Michael Knyszek <mknyszek@google.com>
Commit-Queue: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Cr-Commit-Position: refs/heads/main@{#64109}