[botsregulator] Add tests
original cl: https://chromium-review.googlesource.com/c/infra/infra/+/5392088

Fix: ConsolidateAvailableDUTs implementation uses a map and the test function relies on the ordering of the elements of the map which renders the test flaky.
In this cl we sort the content of resulting slices before comparing them. This renders the test deterministic.

TEST=go test -count=100 ./...; ./botsregulator
BUG=b:325650128

Change-Id: I735d7549f09950e734cc6432ae66dee9f975acdf
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5404270
Reviewed-by: Bryan Ni <bniche@google.com>
Commit-Queue: JB Tamas-Leloup <jtamasleloup@google.com>
Cr-Commit-Position: refs/heads/main@{#64114}