Remove jackneus@google.com from go/src/infra/cros/cmd/prototype-tlw/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/jackneus@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Id0b42771f46d7579fc745d3a2ac09ad2cb9e2d4e