Remove jackneus@google.com from go/src/infra/cros/cmd/prototype-rts/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/jackneus@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I9b13ef8ffe0fe33581cbdd6f508fa227e7c47a10
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5413642
Reviewed-by: Julio Hurtado <juahurta@google.com>
Commit-Queue: Julio Hurtado <juahurta@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#64201}