Remove jackneus@google.com from go/src/infra/cmd/skylab/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/jackneus@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ia3f463101d1ed9de2d0fa4ac5cbb12681107c24d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5414403
Commit-Queue: Andrew Lamb <andrewlamb@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#64204}