satlab: dhcp: remove the IP reservation feature
It was discovered that the ping check used by default by dnsmasq DHCP
server is actually the root cause of the issue with changing IPs of DUTs
connected to Satlab. Changing the approach as it will be way easier to
provide to Satlab users. Leaving only the DNS cleanup.

BUG=b/321300259
TEST=manual

Change-Id: I6bff8f5fc86006c62ebeff8a7ca48ec20f254f10
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5417633
Commit-Queue: Tomasz Jakubowski <jakubowskit@google.com>
Reviewed-by: Michal Matyjek <mmatyjek@google.com>
Cr-Commit-Position: refs/heads/main@{#64222}