recovery: Run modem detection on DUT_POOL_QUOTA
In order to run some cellular tests on DUTs on the QUOTA pool, we first
need to populate the cellular_modem_state label to filter out
cellular-capable DUTs. This creates an additional celluar audit action
which runs on the quota pool, to minimize the overhead on the pool,
no recovery actions are taken if the task fails and the timeouts are
reduced significantly from 2m -> 15s.

BUG=b:331941924
TEST=go make && go test
TEST=run mallet with custom config on quota DUT

Change-Id: I1da78c7a9545c6992963d63e4fdff1cbe51224d1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5421916
Commit-Queue: Jason Stanko <jstanko@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#64255}