Have tryjobs consider platform_re when computing what to rebuild.
This should reduce wasted work when dependencies exist only on
some platforms. However, it's still not completely optimal because
it doesn't distinguish between build and target platforms.

This also does not change the fact that unmodified packages are still
rebuilt if they are dependencies of an affected package -- that is,
production 3pp packages are still not used in this situation whereas
they could be.

Change-Id: I9aad2a13c0558aae01a3d90876031295d7d499d9