Remove sfetisov@google.com from go/src/infra/cmd/drone-prober/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/sfetisov@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ic56f19d39fd91ccb4459f9b88cc55bbc1bc471ba
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5439063
Commit-Queue: Bryan Ni <bniche@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Bryan Ni <bniche@google.com>
Cr-Commit-Position: refs/heads/main@{#64340}