Remove sfetisov@google.com from go/src/infra/cros/botsregulator/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/sfetisov@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I4d34823b1d7a07950225120bf7e8c05eb2c40f6d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5438887
Reviewed-by: Ashish Gandhi <ashishgandhi@google.com>
Commit-Queue: Ashish Gandhi <ashishgandhi@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#64343}