recovery: Fix wifirouter sysupgrade
The sysupgrade command is returning a different error code with
newer openwrt images, likely a result of the ssh server upgrade
in the last image. It works, just the handling in repair makes
it seem like it doesn't. This updates the error handling in
the openwrt wifirouter controller to support the new case.

BUG=b:333756360
TEST=cd ~/cr/infra/go/src/infra/cros/recovery \
&& go clean && go build && make gen && go test ./...
TEST=successfully ran repair locally with a sysupgrade on a testbed
with UU6L OpenWrt APs

Change-Id: I56d63040f1d12e351786c59cf2653f8effb398ca
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5443891
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#64388}