recovery: Add execs for enabling initrd
This adds execs for enabling using initrd/initramfs on btpeers. This
allows us to load a small in-memory filesystem prior to mounting the
real rootfs on the Raspberry Pi devices. This is required so that we can
implement in-place OS upgrades on these btpeers as we need to first
shrink the rootfs on the device which cannot be done when the device is
mounted.

BUG=b:318568744
TEST=local repair
TEST=go test

Change-Id: Idc1b080f66641b617a17d7a7e37bf7c6f826e980
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5362325
Reviewed-by: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Cr-Commit-Position: refs/heads/main@{#64433}