Hardcode -fPIC into CFLAGS.
The current method of setting this in the environment is not
actually working; this is a bug in libffi 3.2.1. However, the
upstream fix for this isn't desirable either since setting custom
CFLAGS causes all of the default optimization flags to be dropped.

Patch the m4 to add -fPIC explicitly as a simpler solution.

linux-riscv64 is on a newer libffi, which is currently building a
completely unoptimized library due to the upstream issue mentioned
above. Patch to add -fPIC for this platform as well, and stop
overriding the default optimization flags.

Change-Id: I2c1f642c8abf3735fd1f94c68d1aa3ac7ea1c1a0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5446249
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/main@{#64492}