[fleetcost] Rename ".../costserver/models" to ".../costserver/entities"
There are two packages named "models", shown below:

- .../fleetcost/api/models
- .../costserver/models

This collision creates some unnecessary complexity where parts of the code use
"models" to mean the generated proto library and parts of the code use it to
mean the datastore entity library.

Let's make the naming a little more consistent and just use "entities" for
everything.

This name is not perfect. "entities" is a commonly used name for a variable
when interacting with datastore, but it is much, much better than
"fleetcostmodels" or similar.

BUG=331291398
TEST=make test

Change-Id: I404cd9c2a7134647a3cdac73f570421dc22b88e6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5452174
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#64501}