[result_adapter] Improve handling of package-level setup failures.
Failure to run the package's common test setup (e.g. in main_test.go)
will result in go test failing, but no failing test results being
uploaded to ResultDB. In addition, the output piped to stdout is
often out of order owing to the concurrent execution of tests,
making it hard to identify which package or error is at fault.

This change starts uploading test results for the packages, in
addition to the tests inside the package. A failure is reported
for the package only if the package reports a failure and its
constituent tests do not report a failure.

BUG=None
TEST=go test infra/tools/result_adapter

Change-Id: Ic9e9bb43a3b3d4ae98687b74318de827e3d808db
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5446255
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#64525}