kajiya: Generate a unique operation name
It's not correct to use (and reuse) the digest of the action as the
operation name. This would also cause problems later when we implement
sending intermediate updates during execution, as the client might
disconnect and try to later get the state of the pending action using
the digest - which would not uniquely identify the operation, as
multiple executions for the same action hash can happen over the
lifetime of a running Kajiya instance.

Change-Id: Ie1481e4d30a9259ab17aa425afb675a3fe731c15
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5446497
Reviewed-by: Fumitoshi Ukai <ukai@google.com>
Commit-Queue: Philipp Wollermann <philwo@chromium.org>
Auto-Submit: Philipp Wollermann <philwo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#64536}