kajiya: Make actioncache and blobstore more robust against races
On Windows, "write to temporary file and rename to final path" is not
atomic in itself, but if we guard all writes in the actioncache and
blobstore to a path with a singleflight.Group, we can ensure that we
will never try to save the store the same file concurrently. This also
simplifies the code a bit.

I extracted the common piece of write-then-rename into the new atomicio
package.

Change-Id: Ibedca66d25353419e292bb73b04597598ec56802
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5446498
Reviewed-by: Junji Watanabe <jwata@google.com>
Commit-Queue: Philipp Wollermann <philwo@chromium.org>
Auto-Submit: Philipp Wollermann <philwo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#64539}