kajiya: Use singleflight instead of sync.Map for merging requests
Note that before this change, the call to actionCache.put was protected
by the mutex in the sync.Map. In this change, we replace this mechanism
with a dedicated singleflight.Group used by the actionCache's put
method.

Follow-up on crrev.com/c/5117085/comment/89fe742d_961fae69/

Change-Id: Ic05cc1d070f6028239cc53bc9d9d4ed34578cdb8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5335429
Reviewed-by: Fumitoshi Ukai <ukai@google.com>
Commit-Queue: Philipp Wollermann <philwo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#64540}