Remove aravindvasudev@google.com from go/src/infra/tools/git/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/aravindvasudev@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ia65244ea9c7d458ba397a2864e5c1dca1c55b8c2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5451179
Commit-Queue: Gavin Mak <gavinmak@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Gavin Mak <gavinmak@google.com>
Cr-Commit-Position: refs/heads/main@{#64547}