Remove aravindvasudev@google.com from recipes/recipe_modules/codesearch/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/aravindvasudev@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I217409f534e6b397621ee5cc265544ffd8b469d4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5451181
Commit-Queue: Gavin Mak <gavinmak@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Gavin Mak <gavinmak@google.com>
Cr-Commit-Position: refs/heads/main@{#64548}