recovery: Add execs to shrink btpeer rootfs partition
This adds execs which make it possible to shrink the rootfs partition on
Raspberry Pi devices in the lab. This is a prerequisite to creating
separate B partitions that we will use to enable A/B booting for
in-place Raspberry Pi image updates in the lab.

BUG=b:318568744
TEST=local repair
TEST=go test

Change-Id: I74ac64153d87595ad3cd1e46d3ae03e03e3339ac
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5362326
Commit-Queue: Jason Stanko <jstanko@google.com>
Reviewed-by: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#64550}