Revert changes to opentelemetry in go.mod
Bug: b/334177235

This reverts the updates to opentelemetry deps made in crrev/c/5384685.
The updated versions seem to cause a memory leak in UFS service. We are
not aware of the root cause. But this change fixes the version of the
libs until a fix is found. See bug for more details

Change-Id: I44affc82dc69e70d32502672b8ad11e95ba3ac00
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5459222
Reviewed-by: Julio Hurtado <juahurta@google.com>
Reviewed-by: Azizur Rahman <azrahman@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#64572}