Remove sfetisov@google.com from go/src/infra/cmd/drone-agent/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/sfetisov@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I877eda6ebae04a7ff92cccbec68087478dbd7021
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5439882
Reviewed-by: Allen Li <ayatane@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Allen Li <ayatane@chromium.org>
Cr-Commit-Position: refs/heads/main@{#64601}