[fleetcost] Use board, model, sku, type, location as key.
Use the board, model, sky, type, and location as keys.

For the type and location, which are enums, just use the number in base 10.

For the other values, which are inherently strings, use the lex64 encoding
(which does not include ; in its output alphabet) to make collisions impossible.

This makes ; totally safe to use as a delimiter.

BUG=331291398
TEST=make test

Change-Id: I650bcd229a594fccaee04cf1c1408cfdfc505676
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5461213
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#64604}