[fleetcost] Show possible values for enum flags (costCadence, indicatorType, and Location)
This CL resolves comment #3 in b/335710807.

Here is a sample of what the error output looks like if you provide a nonexistent cost cadence.

Note that, as of right now, the *entire* usage string will be printed as well in this case and that's unavoidable
given the behavior of subcommands for the time being. Making subcommands more user-friendly is a big undertaking.

```
./fleetcost create-ci -dev -cadence=a |& head -n 10
invalid value "a" for flag -cadence: Choose a candidate from the following values (with or without prefix):
- COST_CADENCE_ANNUALLY
- COST_CADENCE_DAILY
- COST_CADENCE_HOURLY
- COST_CADENCE_MONTHLY
- COST_CADENCE_ONE_TIME
- COST_CADENCE_UNKNOWN
Create a cost indicator

usage:  fleet cost create-ci [options...]
```

BUG=335710807
TEST=make test

Change-Id: I726ecc12bbf0492187df7412c3bc6a6a3966c1ee
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5468565
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#64691}