[golangbuild] respect existing VERSION files from releases
This change modifies the way VERSION files are created when generating
new toolchains. Currently, golangbuild always overwrites the VERSION
file, however this results in a version that respect certain invariants
that downstream tools want to rely on (e.g. that the toolchain version
sorts correctly if the toolchain is mostly just go1.X.Y, perhaps with a
few extra commits).

The semantics after this change are as follows:
- If the input property "version_file" is present, it always overwrites
  the VERSION file with that value.
- If no VERSION file is present or the VERSION file is empty, then the
  VERSION file is written with contents `devel <commit>` or
  `devel <change>/<patchset>` (existing behavior).
- If a VERSION file is present AND the first line matches `go1.X.Y`,
  then only the first line is kept, and we append `-<commit>` or
  `-<change>_<patchset>` to the version.
- If a VERSION file is present otherwise, it is left alone.

Change-Id: I162834ac978cd5068f93265f12ecc2d1a89f3072
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5472883
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Cr-Commit-Position: refs/heads/main@{#64706}