[fleetcost] Use more informative error statuses.
Right now, the fleet cost service by default uses codes.Internal for all of its
errors.  Let's not do that. Let's return a semantically correct error code
instead.

We add a small library that adds a code to an error precisely when one does not
exist already.

This allows us to use codes.Aborted most frequently, but use codes.NotFound or
codes.AlreadyExist in cases where the user is referring to a resource that
already exists or isn't found, respectively.

Here's a sample showing how this works:

$ ./fleetcost get-cost -local -name woof
...

And on the server side (locally):

[I2024-04-19T20:52:20.370165-07:00 83974 0 server.go:2144] 404 POST "http://localhost:8800/prpc/fleetcost.api.rpc.FleetCost/GetCostResult" (0.181783s)

BUG=335947563
TEST=make test

Change-Id: I39ffd06e57354764b2872994de90d7b27b3c282a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5469534
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#64708}