[fleetcost] Create the BigQuery schema for saved fleet cost results.
Results are by DUT, so we report the name.

The name is the hostname or the device id as appropriate.

Known issues:
1) we don't have an explicit time field.

I want this table to be partitioned by ingestion time. I would ideally
prefer this ingestion time field to be present in the proto schema for
explicitness, but I don't know how to make that work with
BQSchemaUpdater yet.

2) Just in general, I want more information on what the dashboards
that will be consuming this table are likely to look like.

BUG=321105751
TEST=make gen

Change-Id: Id07a79128aea481998c76880f06d43abbf4494b9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5477881
Reviewed-by: Vineet Joshi <vkjoshi@google.com>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#64768}