[fleetcost] Remove unnecessary newTransaction parameter.
We can tell from the current context whether a transaction is active
or not. Using the context is the correct way to decide whether to use
a transaction or not, not passing in a new parameter that determines
whether to create a new transaction.

BUG=335710807
TEST=make test

Change-Id: I4826b7888ba64cd3bc7a117e549aeea5449d2cc6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5492522
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#64811}