Add android toolchain CIPD packages
In crbug/40913000, the team stripped down the Android NDK packages to
include only the sysroot, and not the full toolchain. This makes sense
for Chromium, but other projects (such as BoringSSL) are still reliant
on an NDK distribution which contains the full toolchain. As such,
I'd like to add an infra CIPD package for both android_toolchain and
android_toolchain_canary.

Bug: b/334960529
Change-Id: Idca067e93a93c6138c691b8811f41c548f146ab7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5472589
Reviewed-by: Brian Ryner <bryner@google.com>
Commit-Queue: Aaron Knobloch <aknobloch@google.com>
Cr-Commit-Position: refs/heads/main@{#64854}