config: Increase btpeer image download timeout
The decompression timeout in the download exec has been increased
by 5m. This increases the timeout in the config that calls this exec
from 16m to 25m to give plenty of time for the download (10m timeout)
and the decompression (now 10m timeout) plus some extra.

BUG=b:336878507
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && make trees

Change-Id: I651e8f67046e698733a1e985eeb6e453f5c54bfc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5501774
Reviewed-by: Otabek Kasimov <otabek@google.com>
Reviewed-by: Jason Stanko <jstanko@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Cr-Commit-Position: refs/heads/main@{#64934}