[infralib] Add owners to the lib directory of infra.
I want a way to manage common functionality that's shared between a
couple of infra projects, but for which luci/common or similar would
be an inappropriate home.

This libs area is sort of already used for that, but it isn't
maintained all that well and is kind of a random collection of
different utilities.

BUG=None
TEST=owners check that runs as part of presubmit

Change-Id: I28f250586b9b45989d3aefe55abe3e9ab33cbeae
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5503014
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Anushruth Sasivehalli <anushruth@chromium.org>
Reviewed-by: Scott Lee <ddoman@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/main@{#64964}