skylab: Sync WifiRouterFeature enum update for GCMP
A new entry to the WifiRouterFeature enum has been added in
chromiumos/src/config repo for the GCMP feature. This updates
the copy of this enum in the skylab protos to match so that the
new entry is supported in scheduling labels.

BUG=b:337103884
TEST=cd ~/cr/infra/go/src/infra/libs/skylab/inventory && \
go generate && \
go test ./...

Change-Id: I3535b56aeb92237f8c854ec55b358b7bc0715416
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5503027
Commit-Queue: Ruth Mekonnen <rmekonnen@google.com>
Reviewed-by: Jason Stanko <jstanko@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#64975}