Remove denizkara@google.com from go/src/infra/cmd/vmlab/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/denizkara@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ica52a59713e01322fe3a26b30247774d2ffcd2f2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5506873
Commit-Queue: Ming Kong <mingkong@google.com>
Reviewed-by: Ming Kong <mingkong@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#64982}