Remove aravindvasudev@google.com from infra/config/subprojects/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/aravindvasudev@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I016db7eeafe9db2db30cfbbe91e15395a0e1b82d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5451180
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Josip Sokcevic <sokcevic@chromium.org>
Commit-Queue: Josip Sokcevic <sokcevic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#65061}