[infralib] Remove dependency on godebug.pretty in cros_test_platform
The package github.com/kylelemons/godebug/pretty duplicates the
functionality of cmp.Diff and its wrapper,
typed.Got(...).Want(...).Diff().

This seems like as good an opportunity as any to drop the dependency on
a third-party package, and the cost is low, so let's do it.

BUG=338430627
TEST=make test

Change-Id: I7edfc0792ace45deb23122e967137e1848c3f8f7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5512102
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Chris DeLaGarza <cdelagarza@google.com>
Cr-Commit-Position: refs/heads/main@{#65069}