[infralib] Drop dependency on godebug.pretty in cmd/skylab and libs/
The package github.com/kylelemons/godebug/pretty duplicates the
functionality of cmp.Diff and its wrapper,
typed.Got(...).Want(...).Diff().

This seems like as good an opportunity as any to drop the dependency on
a third-party package, and the cost is low, so let's do it.

BUG=338430627
TEST=make test

Change-Id: Ie0ea7ce2369748f99c73d4ec503f30a6bad99882
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5512103
Reviewed-by: Xixuan Wu <xixuan@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#65084}