Remove jaquesc@google.com from go/src/infra/libs/skylab/gs/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/jaquesc@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I4a537f990f334134ccc120e69a7417bab37ec10b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5516158
Reviewed-by: Jaques Clapauch <jaquesc@google.com>
Commit-Queue: Jaques Clapauch <jaquesc@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#65112}