Remove jaquesc@google.com from go/src/infra/tools/dirmd/proto/chromeos/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/jaquesc@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I50c633def56d93049564271cc3a0ebf06408e15a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5515196
Commit-Queue: Jaques Clapauch <jaquesc@google.com>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#65114}