recovery: Always use latest openwrt release
This was technically doing this already, but the 'large' release
number was just hardcoded to 10 and we hadn't gone passed 10 in
the releases so it hadn't been an issue. This increases it to a
very large number and allows setting the version via exec arg.

BUG=None
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && make build
TEST=cd ~/cr/infra/go/src/infra && golangci-lint run \
cros/recovery/internal/execs/wifirouter/...

Change-Id: If519d3d8350ed3c6193693ff80b19a1b27096f2e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5516138
Auto-Submit: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Reviewed-by: Jason Stanko <jstanko@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Cr-Commit-Position: refs/heads/main@{#65120}