Remove jackneus@google.com from go/src/infra/cros/cmd/try/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/jackneus@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I5176019a14eab47d9d5f56f081299b458236276a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5413643
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: George Engelbrecht <engeg@google.com>
Commit-Queue: George Engelbrecht <engeg@google.com>
Cr-Commit-Position: refs/heads/main@{#65181}