[botsregulator] Compute next migration state
This cl introduces the calculation of the next migration state. With a current Drone to CloudBots ratio for a specific model, we calculate the next state based on a target percentage of CloudBots DUTs wanted. This percentage is itself defined by a different set of priorities defined in the migration file. The return value is 2 slices of DUTs to be converted to either Drone or CloudBots.

TEST=go test ./...
BUG=b:334892034

Change-Id: I1992d0993b852ba89d9d778feb2c508f7dae9da2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5512222
Reviewed-by: Bryan Ni <bniche@google.com>
Commit-Queue: JB Tamas-Leloup <jtamasleloup@google.com>
Cr-Commit-Position: refs/heads/main@{#65185}