Remove yuanjunh@google.com from infra/config/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/yuanjunh@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ibfab708c72bd862f15e1699878dbacc9bdb6019e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5523041
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Cr-Commit-Position: refs/heads/main@{#65190}