cros try: add commit to retry
When retrying a build with a buildset tag, `cros try retry` was skipping
that buildset. This commit adds the `commit` flag to the bb add command
so that buildset can be passed through.

In the process, I learned that the way we were unmarshalling json to
protos had a bug in it that was making this very hard to debug.
Essentially, `json.Unmarshal` doesn't support protos the right way, and
so while it's parsing left to right, it parses anything it is capable of
processing until it hits something it chokes on, and then throw an
exception (which we swallow - hooray!). Instead, switch to
`protojson.Unmarshal` that is capable of handling protobufs better, and
thus GitilesCommit is populated.

BUG=b:339007219
TEST=go run . retry --bbid 8748423034391179281 --dryrun

Change-Id: I1d0842d1f78ed2dd16c2ee1ea30c7701a35180dc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5529954
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Reviewed-by: Jard Loucks <jaredloucks@google.com>
Commit-Queue: Benjamin Shai <bshai@google.com>
Cr-Commit-Position: refs/heads/main@{#65337}