Remove sfetisov@google.com from go/src/infra/cros/cmd/labservice/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/sfetisov@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I3d327088c14d2d0e79b44a913aee7c1bfaf898f5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5439064
Commit-Queue: Allen Li <ayatane@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Allen Li <ayatane@google.com>
Cr-Commit-Position: refs/heads/main@{#65408}