Remove et@chromium.org from go/src/infra/appengine/crosskylabadmin/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/et@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I2f82efa20ad740dd9d35e9f73c35515327a1944c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5540109
Reviewed-by: Allen Li <ayatane@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Allen Li <ayatane@google.com>
Cr-Commit-Position: refs/heads/main@{#65409}