kron: refactor the run subcommand
In this large change we refactor the entirety of the buisiness logic in
the run subcommand to be built like a command design pattern. This will
force a common interface for all commands wanting to be implemented in
the future.

In order to accomplish this I de-nested some of the variable types in to
more generalized schemas. I also went through the bb scheduling logic
and reduced the amount of business logic happening at schedule time so
that the caller will be responsible for transforming information to the
correct types

BUG=b:336346974,b:338450514,b:319273179, b:305297604
TEST=unit tests, local CLI runs

Change-Id: I4a3c98d40b149abc958af210544cf6321ede8218
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5517621
Reviewed-by: Varun Srivastav <varunsrivastav@google.com>
Commit-Queue: Julio Hurtado <juahurta@google.com>
Reviewed-by: Azizur Rahman <azrahman@google.com>
Cr-Commit-Position: refs/heads/main@{#65452}