[golangbuild] push the result of make.bash as a test result
This change adds cmdStepTest which can be used to run a command and push
its result to ResultDB as a test result. The implementation is a bit
hacky, but it works.

This change then replaces cmdStepRun for building the toolchain in getGo
with cmdStepTest. Plumbing all the right information through requires
some refactoring, like more buildSpec information. Instead of plumbing
all that manually, this change just passes the buildSpec to getGo. It
also gets rid of getGoFromSpec since the difference is now very small
from getGo. It makes the callsites a little more boilerplate-y, but it's
clearer where Go is being installed in each case.

The motivation behind this change is to push more information into the
test results view in the UI.

While we're here, let's also make variable scoping a bit less
error-prone in the cmd* functions. Currently, it's not super obvious
whether the err from cmdStartStep is shadowing the return variable or
not, making it unclear what's captured by the defer for ending the step.
Name the return variable differently to make that more clear.

For go.dev/issue/67165.

Change-Id: Iec52fba5bf7a5fbd90ad4c286b9f5cdd4862a16c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5551107
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Cr-Commit-Position: refs/heads/main@{#65589}