[shivas] Give DoList and BatchList a channel where they can send responses as they are gathered.
Add a channel of proto Messages to DoList and the other concurrent get
functions that call UFS's paginating APIs.

Adding a channel makes it possible to do work as soon as the protos are
received just by listening on that channel in another goroutine (or
using a helper like dispatcher.Channel).

Please note that I am not taking this opportunity to group the
parameters to DoList(...) and friends into a struct or use the option
passing idiom. These functions are acquiring one parameter every four
years or so by my estimation. That growth is sustainable. Today is not
the day for change.

BUG=321105253
TEST=./shivas get dut AND observe results

Change-Id: Ida5fb5348104499e9f934c30a9eca102b5c9f1a4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5560157
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Reviewed-by: Justin Suen <justinsuen@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#65665}