[golangbuild] log raw JSON from Go tests
This change adds logging for the raw JSON emitted by Go tests for
debugging and understanding why the output appears the way that it does.
Questions around why the output contains some things and not others
would be much easier to answer with the raw JSON to compare it to.

Change-Id: Iff76163082ada389e9cfbf48a6ed0a84a6ebc824
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5560486
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#65713}