[fleetcost] Add readonly flag to the persist-to-bigquery RPC
Add a readonly flag that will prevent actual writes to bigquery.

We want the default, at the RPC level, to be to WRITE to bigquery
because, realistically, the main use case for a readonly flag is
development. That's an important use case, but not important enough to
be the default.

BUG=321105253
TEST=make test

Change-Id: I78ab4b8721484db8807a6d0aeb52b2fa829ed1e9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5575697
Reviewed-by: Anushruth Sasivehalli <anushruth@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#65767}