skylab: Sync WifiRouterFeature enum update for SAE-EXT-KEY
A new entry to the WifiRouterFeature enum has been added in
chromiumos/src/config repo for the SAE-EXT-KEY feature which is
necessary for AMK-24/25 tests. This updates the copy of this enum
in the skylab protos to match so that the new entry is supported
in scheduling labels.

BUG=b:339435290
TEST=cd ~/chrome_infra/infra/go/src/infra/libs/skylab/inventory &&
go generate && go test ./...

Change-Id: Ia54541ea7bb0f058ae117757aaf9dde3ef0ce615
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5530991
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Ruth Mekonnen <rmekonnen@google.com>
Reviewed-by: Jason Stanko <jstanko@google.com>
Cr-Commit-Position: refs/heads/main@{#65773}